requirements.txt
# Core dependencies
numpy==1.20.3
pandas==1.3.4
scipy==1.7.1
six==1.16.0
tqdm==4.62.3
PyYAML==6.0
Pillow==9.4.0
pyarrow
decord

# ML and visualization
scikit-learn
matplotlib==3.4.3
pydot
pydotplus
graphviz

# OpenCV & MediaPipe
opencv-python
mediapipe==0.10.11
